feat(security): distribute serverless functions across least-privilege service accounts + api_proxy role
Implements issue #1761 across four phases:

Phase 0 – Shared dispatcher harness
- Add functions/shared_libs/node/dispatcher.cjs with createDispatcherLogger,
  validateSecret, formatResponse, redactForLogging (console fallback when
  winston unavailable outside function packages)

Phase 1 – Stripe functions isolated to nodePayments Cloud Function
- Create functions/nodePayments/ (nodejs22, port 42021, entry point nodePayments)
  routing createStripeCheckout, createStripeBasePrice, saveAddonMappings
- Remove stripe dep from callNodeFunction/package.json
- Repoint actions.yaml handler URLs to CLOUD_FUNCTION_LINK_NODE_PAYMENTS

Phase 2 – api_proxy read-only Hasura role
- Add SELECT-only api_proxy role to 12 tables: Course, Program, Session,
  CourseLocation, CourseDefaultAddresses, CourseGroup, CourseGroupOption,
  CourseFundingOrganization, Organization, CourseEnrollment, Attendance, User
- Remove createCertificate Hasura action (dead code)
- Set HASURA_USE_READ_ONLY_ROLE=true on api-proxy Cloud Function in Terraform
- EduHubClient reads HASURA_USE_READ_ONLY_ROLE env var and sets x-hasura-role header

Phase 3 – Dead-code removal
- Delete functions/callNodeFunction/createCertificate/ and its Hasura action

Phase 4 – Keycloak admin functions merged into keycloakAdmin Cloud Function
- Create functions/keycloakAdmin/ (nodejs22, port 42001) routing addKeycloakRole,
  removeKeycloakRole, updateFromKeycloak via req.headers.name
- Delete standalone addKeycloakRole, removeKeycloakRole, updateFromKeycloak functions
- Update event trigger webhooks in CourseInstructor and OrganizationAdmin YAMLs
- Update updateFromKeycloak action to point to CLOUD_FUNCTION_LINK_KEYCLOAK_ADMIN

Infrastructure
- Replace single shared SA with 7 per-function least-privilege SAs in 06_cloud-functions.tf:
  api_proxy_sa, call_python_function_sa, call_node_function_sa, send_mail_sa,
  send_questionaires_sa, node_payments_sa, keycloak_admin_sa
- Each SA gets only the Secret Manager bindings it needs (principle of least privilege)
- Update 05_hasura.tf: remove old env vars, add CLOUD_FUNCTION_LINK_NODE_PAYMENTS
  and CLOUD_FUNCTION_LINK_KEYCLOAK_ADMIN
- Update CI workflow to zip nodePayments and keycloakAdmin, drop retired functions
- Update docker-compose.yml with new function link env vars
