+++ backend/metadata/actions.graphql
@@ -19,14 +19,6 @@ type CopyProjectFromTemplateResult {
 }
 
 type Mutation {
-  createCertificate(
-    userIds: [uuid!]!
-    courseId: Int!
-    certificateType: String!
-  ): result
-}
-
-type Mutation {
   createCertificates(
     userIds: [uuid!]!
     courseId: Int!
