+++ backend/metadata/databases/default/tables/public_CourseFundingOrganization.yaml
@@ -9,6 +9,13 @@ object_relationships:
     using:
       foreign_key_constraint_on: organizationId
 select_permissions:
+  - role: api_proxy
+    permission:
+      columns:
+        - id
+        - courseId
+        - organizationId
+      filter: {}
   - role: anonymous
     permission:
       columns:
