+++ backend/metadata/databases/default/tables/public_CourseInstructor.yaml
@@ -59,10 +59,12 @@ event_triggers:
       interval_sec: 10
       num_retries: 0
       timeout_sec: 60
-    webhook_from_env: CLOUD_FUNCTION_LINK_ADD_KEYCLOAK_ROLE
+    webhook_from_env: CLOUD_FUNCTION_LINK_KEYCLOAK_ADMIN
     headers:
       - name: secret
         value_from_env: HASURA_CLOUD_FUNCTION_SECRET
+      - name: name
+        value: addKeycloakRole
       - name: role
         value: instructor
     cleanup_config:
