+++ backend/metadata/databases/default/tables/public_User.yaml
@@ -93,6 +93,14 @@ array_relationships:
           name: Project
           schema: public
 select_permissions:
+  - role: api_proxy
+    permission:
+      columns:
+        - id
+        - firstName
+        - lastName
+        - email
+      filter: {}
   - role: anonymous
     permission:
       columns:
