+++ docker-compose.yml
@@ -133,13 +133,11 @@ services:
       HASURA_GRAPHQL_JWT_SECRET: '{"jwk_url": "http://keycloak:8080/realms/edu-hub/protocol/openid-connect/certs"}'
 
       HASURA_BUCKET: emulated-bucket
-      CLOUD_FUNCTION_LINK_UPDATE_FROM_KEYCLOAK: "http://node_functions:42020"
       CLOUD_FUNCTION_LINK_SEND_MAIL: "http://node_functions:42019"
-      CLOUD_FUNCTION_LINK_ADD_KEYCLOAK_ROLE: "http://node_functions:42000"
-      CLOUD_FUNCTION_LINK_REMOVE_KEYCLOAK_ROLE: "http://node_functions:42002"
-      CLOUD_FUNCTION_LINK_CREATE_PARTICIPATION_CERTIFICATE: "http://node_functions:42023"
       CLOUD_FUNCTION_LINK_SEND_QUESTIONAIRES: "http://node_functions:42024"
       CLOUD_FUNCTION_LINK_CALL_NODE_FUNCTION: "http://node_functions:42022"
+      CLOUD_FUNCTION_LINK_NODE_PAYMENTS: "http://node_functions:42021"
+      CLOUD_FUNCTION_LINK_KEYCLOAK_ADMIN: "http://node_functions:42001"
       CLOUD_FUNCTION_LINK_CALL_PYTHON_FUNCTION: "http://python_functions:42025"
     volumes:
       - ./backend:/hasura
