+++ functions/nodePayments/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "nodePayments",
+  "version": "0.1.0",
+  "type": "module",
+  "dependencies": {
+    "graphql-request": "^6.1.0",
+    "stripe": "^14.21.0",
+    "winston": "^3.3.3"
+  },
+  "devDependencies": {
+    "@google-cloud/functions-framework": "^3.1.2",
+    "npm-watch": "^0.13.0"
+  },
+  "watch": {
+    "start": "*.js"
+  },
+  "scripts": {
+    "watch": "npm-watch start",
+    "start": "npx @google-cloud/functions-framework --port 42021 --target=nodePayments"
+  }
+}
