--- functions/removeKeycloakRole/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "removeKeycloakRole",
-  "version": "0.1.0",
-  "type": "module",
-  "dependencies": {
-    "@keycloak/keycloak-admin-client": "20.0.0",
-    "body-parser": "1.20.3",
-    "node-fetch": "2.6.7",
-    "graphqurl": "1.0.1"
-  },
-  "devDependencies": {
-    "npm-watch": "^0.6.0",
-    "@google-cloud/functions-framework": "^3.1.2"
-  },
-  "watch": {
-    "start": "*.js"
-  },
-  "scripts": {
-    "watch": "npm-watch start",
-    "start": "npx @google-cloud/functions-framework --port 42002 --target=removeKeycloakRole"
-  }
-}
