--- functions/updateFromKeycloak/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "updateFromKeycloak",
-  "version": "0.1.0",
-  "dependencies": {
-    "@keycloak/keycloak-admin-client": "^16.1.0",
-    "body-parser": "1.20.3",
-    "node-fetch": "2.6.7",
-    "graphqurl": "1.0.1"
-  },
-  "devDependencies": {
-    "npm-watch": "^0.6.0",
-    "@google-cloud/functions-framework": "^3.1.2"
-  },
-  "watch": {
-    "start": "*.js"
-  },
-  "scripts": {
-    "watch": "npm-watch start",
-    "start": "npx @google-cloud/functions-framework --port 42020 --target=updateFromKeycloak"
-  }
-}
