+++ infrastructure/application/05_hasura.tf
@@ -86,24 +86,20 @@ module "hasura_service" {
       value = google_cloudfunctions2_function.call_python_function.service_config[0].uri
     },
     {
-      name  = "CLOUD_FUNCTION_LINK_ADD_KEYCLOAK_ROLE"
-      value = google_cloudfunctions2_function.add_keycloak_role.service_config[0].uri
-    },
-    {
-      name  = "CLOUD_FUNCTION_LINK_REMOVE_KEYCLOAK_ROLE"
-      value = google_cloudfunctions2_function.remove_keycloak_role.service_config[0].uri
-    },
-    {
-      name  = "CLOUD_FUNCTION_LINK_UPDATE_FROM_KEYCLOAK"
-      value = google_cloudfunctions2_function.update_from_keycloak.service_config[0].uri
-    },
-    {
       name  = "CLOUD_FUNCTION_LINK_SEND_MAIL"
       value = google_cloudfunctions2_function.send_mail.service_config[0].uri
     },
     {
       name  = "CLOUD_FUNCTION_LINK_CALL_NODE_FUNCTION"
       value = google_cloudfunctions2_function.call_node_function.service_config[0].uri
+    },
+    {
+      name  = "CLOUD_FUNCTION_LINK_NODE_PAYMENTS"
+      value = google_cloudfunctions2_function.node_payments.service_config[0].uri
+    },
+    {
+      name  = "CLOUD_FUNCTION_LINK_KEYCLOAK_ADMIN"
+      value = google_cloudfunctions2_function.keycloak_admin.service_config[0].uri
     },
     {
       name  = "CLOUD_FUNCTION_LINK_SEND_QUESTIONAIRES"
